Make argument n optional for pounders. Close #20.
The argument n that determines the number of residuals for pounders is now optional. If n is unset, the R code will evaluate the objective function once to determine n. If n is set, it will just take the number and assume it's correct. Updated README, documentation, and tests correspondingly.